xml schema checker is implemented

+++ SmallTuba/SmallTuba/Utility/TimeConverter.cs
@@ -18,7 +18,7 @@ namespace SmallTuba.Utility
         }
 
 
-        public static double ConvertToUnixTimestamp(DateTime date)
+        public static int ConvertToUnixTimestamp(DateTime date)
         {
             DateTime origin = new DateTime(1970, 1, 1, 0, 0, 0, 0);
             TimeSpan diff = date - origin;
